+++ Changelog.md
@@ -1,4 +1,6 @@
 ## Change Log
+- Next
+  - Add ElvUI/Tukui skin support
 - Version 2.2.8 - 2024/08/20
   - Add Warband internal balance (fix total balance calculation)
   - Hide characters and realms if cashless